Fix an error where the image link is invalid
Discord threw this 👇
discord.errors.HTTPException: BAD REQUEST (status code: 400): Not a well formed URL.
Randomly in my Google search command. So I "investigated" it and found that the "cse_image" for certain searches has a URL of the sort "x-raw-image:///<stuff>". Discord obviously doesn't recognise this as an image url. Nor did I ever see such an url. So I opened up the json response of cse for that specific search and tried to find a solution. The only one I found was the "cse_thumbnail" attribute. (Which isn't as high quality but is the only alternative) . There was of course another image url in the json, "og:image". But its url was broken.

So here's a tiny fix for the error. 
Also, you can reproduce the error by searching "free code camp python" and it's in the first result.